build: add workflows + move documentation to github pages as jekyll/liquid project with custom build
