+++ .github/workflows/github_pages.yml
@@ -1,0 +1,55 @@
+name: Deploy Jekyll site to Pages
+on:
+    push:
+        tags:
+            - "*.*.*"
+            - "!*.*.*.beta.*"
+    workflow_dispatch:
+permissions:
+    contents: read
+    pages: write
+    id-token: write
+concurrency:
+    group: "pages"
+    cancel-in-progress: false
+jobs:
+    build:
+        runs-on: ubuntu-latest
+        steps:
+            - name: Checkout
+              uses: actions/checkout@v4
+            - name: Install
+              run: npm ci
+            - name: Generate types.json
+              run: npm run typedoc
+            - name: Setup Ruby
+              uses: ruby/setup-ruby@v1
+              with:
+                  ruby-version: "3.4"
+                  bundler-cache: true
+                  cache-version: 0
+            - name: Setup Pages
+              id: pages
+              uses: actions/configure-pages@v5
+            - name: Change to Jekyll Dir
+              run: cd documentation
+            - name: Install Ruby Gems
+              run: bundle install
+              env:
+                  JEKYLL_ENV: production
+            - name: Build with Jekyll
+              run: bundle exec jekyll build --baseurl "${{ steps.pages.outputs.base_path }}"
+              env:
+                  JEKYLL_ENV: production
+            - name: Upload artifact
+              uses: actions/upload-pages-artifact@v3
+    deploy:
+        environment:
+            name: github-pages
+            url: ${{ steps.deployment.outputs.page_url }}
+        runs-on: ubuntu-latest
+        needs: build
+        steps:
+            - name: Deploy to GitHub Pages
+              id: deployment
+              uses: actions/deploy-pages@v4
