+++ documentation/_includes/example.md
@@ -1,0 +1,7 @@
+So valid YAML for this could look like this:
+
+```yaml
+{{ include.options | examplify: include.type | toYAML }}
+```
+
+I used all available options in this example. You don`t need to do this! Take a look at the required fields for what you need!
