+++ documentation/_includes/label-sort.md
@@ -1,0 +1,8 @@
+{: .note }
+>If you want to sort anything in the strategies, you can use "Label Sort".
+>
+>It works currently on areas (=sort of navigation areas and views overall) and entities (=sort in the grid).
+>
+>Just create Labels with the exact name as below (Sort:1, Sort:2, ... as many as you want) and assign them.
+>
+>![Label Sort](/assets/area/area-strategy-label-sort.png "Label Sort")
