+++ documentation/_includes/rows.md
@@ -1,0 +1,4 @@
+## Content/Rows
+
+Every Row consists of a (optional) title and a Grid with all the cards for the entities.
+The row is defined by:
