+++ documentation/_plugins/ternary.rb
@@ -1,0 +1,9 @@
+module Jekyll
+  module Ternary
+    def ternary (input, truthyValue, falsyValue)
+        input ? truthyValue : falsyValue
+    end
+  end
+end
+
+Liquid::Template.register_filter(Jekyll::Ternary)
