--- src/doc_snippets/area/tab.md
@@ -1,8 +1,0 @@
-## Tabs
-
-The tabs shown in the the main section.
-
-Define as many Tabs as you want.
-The Tab will only be shown in the View per Area when it has content.
-
-Example: if the Living Room has a Tab where none of the rows would have entities (like Camera, which has only one row and can easily be empty) the whole Tab is hidden.
