--- src/doc_snippets/battery/default-config.md
@@ -1,13 +1,0 @@
-# Default Config explained
-
-The default config is written as yaml. This should make it easy to make your own changes because it is written exactly as you would in your dashboard.
-
-You can find it [here](/src/config/gridDefaultConfig.yml)
-
-Additionally for this View Strategy the platforms options is hard-coded so it is not defined in yaml but can be overwritten!
-
-There are Rows for:
-
-- Zigbee
-- Switchbot
-- Other (always present everything that is not defined in platforms!)
