--- src/doc_snippets/common/example.md
@@ -1,7 +1,0 @@
-So a valid configuration could look like this:
-
-```yaml
-{{ example | toYAML }}
-```
-
-I used all available options in this example. You don`t need to do this!
