--- src/doc_snippets/common/rows.md
@@ -1,5 +1,0 @@
-## Content/Rows
-
-Define the Rows in the Tab.
-Every Row consists of a (optional) title and a Grid with all the cards for the entities.
-The row is defined by:
