--- src/doc_snippets/log/log-preset.md
@@ -1,5 +1,0 @@
-## Log Preset
-
-Defines the Button on the side and which entities will be loaded in the Log Card when the Preset Button is clicked.
-
-Define as many Log Presets as you want.
