+++ src/index.ts
@@ -1,4 +1,4 @@
-import "./area-strategy";
+import "./area-dashboard-strategy";
 import "./log-view-strategy";
 import "./grid-view-strategy";
 import "./battery-view-strategy";
