+++ typedoc.json
@@ -1,6 +1,6 @@
 {
     "entryPointStrategy": "expand",
-    "json": "docs/typedoc.json",
+    "json": ".typedoc/typedoc.json",
     "readme": "none",
     "blockTags": ["@description", "@defaultValue", "@remarks", "@example"],
     "sort": ["enum-member-source-order", "source-order"]
